adding 1d torch files
