+++ kymatio/scattering1d/__init__.py
@@ -1,4 +1,4 @@
 __all__ = ['Scattering1D']
 
+from .frontend import Scattering1D
 
-from .scattering1d import Scattering1D
